requirements.txt
flask==3.0.3
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
flask-cors==4.0.1
gunicorn==22.0.0
